requirements.txt
fastapi
uvicorn[standard]
youtube-transcript-api
python-dotenv
google-genai
yt-dlp
openai-whisper
ffmpeg-python
